Change int in data record to _32

+++ output/sketch/DataRecord.h
@@ -5,20 +5,20 @@ struct DataRecord
 {
     unsigned long time;
     int32_t measurementTime; // Tijd van meting in EPOCH
-    int16_t P1;
-    int16_t P2;
-    int16_t P3;
-    int16_t P4;
-    int16_t P5;
-    int16_t P6;
-    int16_t P7;
-    int16_t P8;
-    int16_t P9;
-    int16_t P10;
-    int16_t P11;
-    int16_t P12;
-    int16_t P13;
-    int16_t P14;
-    int16_t P15;
-    int16_t P16;
+    int32_t P1;
+    int32_t P2;
+    int32_t P3;
+    int32_t P4;
+    int32_t P5;
+    int32_t P6;
+    int32_t P7;
+    int32_t P8;
+    int32_t P9;
+    int32_t P10;
+    int32_t P11;
+    int32_t P12;
+    int32_t P13;
+    int32_t P14;
+    int32_t P15;
+    int32_t P16;
 };
